+++ src/main/java/com/github/technus/tectech/thing/metaTileEntity/multi/GT_MetaTileEntity_EM_bhg.java
@@ -70,7 +70,7 @@ public class GT_MetaTileEntity_EM_bhg extends GT_MetaTileEntity_MultiblockBase_E
     private static final Block[] blockType = new Block[]{sBlockCasingsTT,sBlockCasingsTT,sBlockCasingsTT,sBlockCasingsTT,sBlockCasingsTT};
     private static final byte[] blockMeta = new byte[]{12, 13, 14, 10, 11};
     private static final String[] addingMethods = new String[]{"addClassicToMachineList", "addElementalToMachineList"};
-    private static final byte[] casingTextures = new byte[]{textureOffset, textureOffset + 4};
+    private static final short[] casingTextures = new short[]{textureOffset, textureOffset + 4};
     private static final Block[] blockTypeFallback = new Block[]{sBlockCasingsTT, sBlockCasingsTT};
     private static final byte[] blockMetaFallback = new byte[]{0, 4};
     //endregion
